mergen/worker/pyproject.toml
[project]
name = "zgrbid-worker"
version = "0.1.0"
description = "Background worker for ZgrBid"
dependencies = [
  "rq>=1.15.0",
  "redis>=5.0.0",
  "python-dotenv>=1.0.0",
  "sqlalchemy>=2.0.0",
  "psycopg[binary]>=3.1.0",
  "pdfplumber>=0.10.0",
  "openpyxl>=3.1.0",
  "sentence-transformers>=2.2.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"



